+++ packages/frontend/vite.config.ts
@@ -1,14 +1,15 @@
 import { resolve } from "path";
 import { config } from "../config";
 
+const target = `http://${config.server.host}:${config.server.port}`;
+
 module.exports = {
   outDir: "../../dist/public",
   alias: {
     "/@/": resolve(__dirname, "./src"),
   },
   proxy: {
-    "/graphql": {
-      target: `http://${config.server.host}:${config.server.port}`,
-    },
+    "/graphql": { target },
+    "/playground": { target },
   },
 };